requirements.txt
# Algorithmic Trading System Requirements

# Web Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0
jinja2==3.1.6
python-multipart==0.0.20
websockets==14.2

# Trading APIs
ibapi==9.81.1.post1
alpha-vantage==3.0.0
polygon-api-client==1.15.4
yfinance==0.2.66

# Data Processing
pandas==2.3.2
numpy==2.3.3
ta-lib==0.6.7

# Machine Learning & Analysis
scikit-learn==1.7.2
matplotlib==3.10.6
plotly==6.3.0
seaborn==0.13.2

# Database & Storage
sqlalchemy==2.0.43
alembic==1.16.5
redis==6.4.0

# Task Queue & Scheduling
celery==5.5.3
schedule==1.2.2

# HTTP & Networking
requests==2.32.5
aiohttp==3.12.15
asyncio-mqtt==0.16.2

# Utilities
python-dotenv==1.1.1
loguru==0.7.3
